Update precommit and fix new issues. Bugfix womd inner reference in toVectorOfVectors. Bugfix womd roadgraph dir should be dx,dy,dz. Added WOMD reading benchmark and simple sample visualization.

+++ include/replay_parsing.hpp
@@ -60,7 +60,7 @@ class UpgradeState
     void setRace(Race race) noexcept;
 
     /**
-     * @brief Calculate and set the expected reserach completion times based on the actions and their timepoint
+     * @brief Calculate and set the expected research completion times based on the actions and their timepoint
      * @param playerActions Vector of actions the player has performed at each timepoint
      * @param gameTime The timepoint in the game each action is associated with
      */
@@ -78,8 +78,9 @@ class UpgradeState
     [[nodiscard]] auto getState(std::size_t timeIdx) const -> py::array_t<T>
     {
         py::array_t<T> state({ static_cast<py::ssize_t>(upgradeTimes_.size()) });
-        std::ranges::transform(
-            upgradeTimes_, state.mutable_data(), [=](int32_t time) { return static_cast<T>(timeIdx > time ? 1 : 0); });
+        std::ranges::transform(upgradeTimes_, state.mutable_data(), [=](int32_t time) {
+            return static_cast<T>(static_cast<int32_t>(timeIdx) > time ? 1 : 0);
+        });
         return state;
     }
